lib: introduce compress_get_hpointer64
Allow to pass unwrapped avail value to applications. Introduce
new compress_get_hpointer64() function and change callback
in ops.

Use UINT_MAX limiting rather than blind wrap like in the previous
code in compress_get_hpointer().

Closes: https://github.com/alsa-project/tinycompress/pull/31

+++ include/tinycompress/compress_ops.h
@@ -23,7 +23,7 @@ struct compress_ops {
 			unsigned int flags, struct compr_config *config);
 	void (*close)(void *compress_data);
 	int (*get_hpointer)(void *compress_data,
-			unsigned int *avail, struct timespec *tstamp);
+			unsigned long long *avail, struct timespec *tstamp);
 	int (*get_tstamp)(void *compress_data,
 			unsigned long long *samples, unsigned int *sampling_rate);
 	int (*write)(void *compress_data, const void *buf, size_t size);
